Parser rewrite
There was a problem parsing function declarations where the function
has a parameter that is a function pointer and the parameter is not
named. Also, apart from this bug I wasn't happy with how the code
worked - it was attempting to parse a declarator, then catching an
exception if that didn't work and then trying to parse an abstract
dedlarator. That was alway supposed to be a stop gap until I could
figure out a better way of doing things. The better way was just to
take the grammar in c23 for declarators and (re)implement a parser for
it.