Fix constant expression bugs
This compiler started out just allowing fixed constants in array
sizes, but then added support for using expressions that evaluate to a
constant in their place, but it didn't always work because we were
trying to evaluate the expression while parsing, and it wasn't safe to
do the before varaible resolution and type checking. So I made a new
type of Constant called ConstantExp, which can be used for the array
size while parsing, and then it's evaluated to a real Constant in type
checking.